Adding _connection_is_secure() to Payment Class and configuration.
To make it simplier for people who may not have their full site forced to an SSL connection, this will redirect the end user if they attempt to access a page which is not secure and the payment class is being loaded.  You can turn off the redirection by setting 'force_secure_connection' to FALSE in the configuration. Then it will only log an error that the page is being used without a secure connection.