Adding the Bluepay Gateway
This is the initial support for the bluepay gateway. It provides the basic
credit card sale transactions such as AUTH, CAPTURE, REFUND, VOID, and of
course the SALE itself. A list of unsupported API parameters is included
inside the configuration file. The other methods to fetch lists of transactions
will be added later.